# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API SIHIOS-TS",
        "description": "API de la plateforme SIHIOS-TS",
        "contact": {
            "name": "Developer",
            "email": "michel.banh@softskills.ci"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/parcours/list": {
            "get": {
                "tags": [
                    "Parcours"
                ],
                "summary": "Obtenir la liste des parcours",
                "operationId": "79c274ddb53050cb8a6ab07d03a2a5e3",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Une liste de parcours récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Parcour"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/parcours/store": {
            "post": {
                "tags": [
                    "Parcours"
                ],
                "summary": "Créer un nouveau parcours",
                "operationId": "ebabf85635d812e353bb30d0e931bb70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Titre du parcours"
                                    },
                                    "field": {
                                        "type": "string",
                                        "example": "Domaine d'étude"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description du parcours"
                                    },
                                    "file": {
                                        "description": "Fichier à télécharger",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parcours créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Parcour"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/parcours/detail{id}": {
            "get": {
                "tags": [
                    "Parcours"
                ],
                "summary": "Obtenir les détails d'un parcours par ID",
                "operationId": "96158bb26d11babf08bbdfadf06c26ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du parcours à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du parcours récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Parcour"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parcours non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/parcours/update/{id}": {
            "put": {
                "tags": [
                    "Parcours"
                ],
                "summary": "Mettre à jour un parcours existant",
                "operationId": "add0796d0330fab44543dbccb8eddbe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du parcours à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Titre mis à jour du parcours"
                                    },
                                    "field": {
                                        "type": "string",
                                        "example": "Domaine d'étude mis à jour"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description mise à jour du parcours"
                                    },
                                    "file": {
                                        "description": "Fichier à télécharger (optionnel)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Parcours mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Parcour"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parcours non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/parcours/delete/{id}": {
            "delete": {
                "tags": [
                    "Parcours"
                ],
                "summary": "Supprimer un parcours par ID",
                "operationId": "b251c65f54978f7b08e6f63ae34f1783",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du parcours à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parcours supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parcours non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Parcour": {
                "title": "Parcours",
                "required": [
                    "id",
                    "label",
                    "field",
                    "file_path",
                    "mime_type",
                    "size"
                ],
                "properties": {
                    "id": {
                        "description": "L'identifiant unique du parcours",
                        "type": "integer"
                    },
                    "label": {
                        "description": "Titre du parcours",
                        "type": "string"
                    },
                    "field": {
                        "description": "Domaine d'étude du parcours",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description du parcours",
                        "type": "string"
                    },
                    "file_path": {
                        "description": "Chemin du fichier téléchargé",
                        "type": "string"
                    },
                    "mime_type": {
                        "description": "Type MIME du fichier téléchargé",
                        "type": "string"
                    },
                    "size": {
                        "description": "Taille du fichier téléchargé en octets",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Utilisez un jeton JWT pour l'authentification.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}